Resilio Connect API/Python3/shotgrid-webhooks-firebase/functions/requirements.txt
functions-framework==3.*
firebase-functions==0.1.*
shotgun_api3
requests
PyYAML
